refactor: move do_mouse() and its helpers to mouse.c
Vim moved it in patch 8.1.2062, and this is required for patch 9.0.0822.
Also change macros in mouse.h to enums.
